resurrect the vdnsbasic test in train
vrouter agent uses vm name to generate vdns entries. in queens it
is explcit nova instance display_name attribute value. but in train
we have an InstanceInfo object without the display name thus vrouter
agent generates vdns entries with rec_name equals to the instance
name which is a name of the libvirt entity and doesn't match the
name the test expects.
let's read the instance entity from db directly and use its display_name
if the read is successful.

Change-Id: Ibce7168af281a84b4c5d2e0fe73fd61ce8712310